+++ PolarAlignment/Properties/Settings.Designer.cs
@@ -574,5 +574,17 @@ namespace NINA.Plugins.PolarAlignment.Properties {
                 this["OAPAHomeY"] = value;
             }
         }
+
+        [global::System.Configuration.UserScopedSettingAttribute()]
+        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
+        [global::System.Configuration.DefaultSettingValueAttribute("5")]
+        public double MaxAutomatedCorrectionMagnitude {
+            get {
+                return ((double)(this["MaxAutomatedCorrectionMagnitude"]));
+            }
+            set {
+                this["MaxAutomatedCorrectionMagnitude"] = value;
+            }
+        }
     }
 }